+++ findempro/simulate/urls.py
@@ -7,5 +7,5 @@ app_name = 'simulate'
 
 urlpatterns = [
     path("init", view=simulate_show_view, name="simulate.show"),
-    path("result", view=simulate_result_simulation_view, name="simulate.result"),
+    path("result/<int:pk>/", view=simulate_result_simulation_view, name="simulate.result"),
 ]
