+++ findempro/simulate/utils.py
@@ -4,64 +4,3 @@ from .models import ResultSimulation
 def get_results_for_simulation(simulation_id):
     return ResultSimulation.objects.filter(fk_simulation_id=simulation_id)
 
-def analyze_simulation_results(results):
-    # Example: Calculate average demand mean and standard deviation
-    avg_demand_mean = sum(result.demand_mean for result in results) / len(results)
-    avg_demand_std_dev = sum(result.demand_std_deviation for result in results) / len(results)
-
-    # You can add more analysis based on your specific requirements
-
-    return {
-        'avg_demand_mean': avg_demand_mean,
-        'avg_demand_std_dev': avg_demand_std_dev
-        # Add more metrics as needed
-    }
-
-def decision_support(analysis_results):
-    # Example: Decision support logic based on average demand mean
-    if analysis_results['avg_demand_mean'] > 100:
-        return "Consider increasing inventory levels for better service."
-    else:
-        return "Inventory levels seem sufficient."
-
-def decision_support_product_diversification(analysis_results):
-    # Add your decision logic here
-    pass
-
-def decision_support_production_cost_optimization(analysis_results):
-    # Add your decision logic here
-    pass
-
-def decision_support_inventory_management(analysis_results):
-    if analysis_results['avg_demand_mean'] > 100:
-        return "Consider increasing inventory levels for better service."
-    else:
-        return "Inventory levels seem sufficient."
-
-def decision_support_pricing_strategies(analysis_results):
-    # Add your decision logic here
-    pass
-
-def decision_support_financial_planning(analysis_results):
-    # Add your decision logic here
-    pass
-
-def decision_support_funding_search(analysis_results):
-    # Add your decision logic here
-    pass
-
-def decision_support_technology_investment(analysis_results):
-    # Add your decision logic here
-    pass
-
-def decision_support_risk_analysis(analysis_results):
-    # Add your decision logic here
-    pass
-
-def decision_support_strategic_alliances_evaluation(analysis_results):
-    # Add your decision logic here
-    pass
-
-def decision_support_customer_relationships_maintenance(analysis_results):
-    # Add your decision logic here
-    pass
